requirements.txt
django
social-auth-app-django
django-phone-field
django-paypal